refactor(access): move ModeCompare's comparison predicates into a module
Pure move, no behaviour change — the bug they carry is fixed in the next
commit, so the fix is three lines rather than three lines inside a
140-line new file.

They were closures inside a `useEffect` in a 759-line dialog, so nothing
could reach them without rendering the whole thing. That is why the defect
in two of them has stood unnoticed. As pure functions of
`(allModes, selectedModeNames, …)` they are testable, and as plain `.ts`
they survive the component's eventual Lit conversion — the markup will not.
